+++ src/main/java/com/zxf/service/CityService.java
@@ -9,7 +9,7 @@ public interface CityService {
      * @param id
      * @return
      */
-    City findCityById(Long id);
+    City findCityById(int id);
 
     /**
      * 新增城市信息
@@ -17,7 +17,7 @@ public interface CityService {
      * @param city
      * @return
      */
-    Long saveCity(City city);
+    int saveCity(City city);
 
     /**
      * 更新城市信息
@@ -25,7 +25,7 @@ public interface CityService {
      * @param city
      * @return
      */
-    Long updateCity(City city);
+    int updateCity(City city);
 
     /**
      * 根据城市 ID,删除城市信息
@@ -33,6 +33,6 @@ public interface CityService {
      * @param id
      * @return
      */
-    Long deleteCity(Long id);
+    int deleteCity(int id);
 
 }
